+++ podcasts/AppDelegate+Analytics.swift
@@ -3,21 +3,30 @@ import PocketCastsUtils
 import PocketCastsDataModel
 
 extension AppDelegate {
-    private var shouldRegisterFirstPartyAdapters: Bool {
-        UIApplication.shared.isProtectedDataAvailable && !Settings.analyticsOptOut() && !Analytics.shared.adaptersRegistered
+    private var shouldRegisterAdapters: Bool {
+        UIApplication.shared.isProtectedDataAvailable && !Analytics.shared.adaptersRegistered
     }
 
     func setupAnalytics() {
+        guard shouldRegisterAdapters else {
+            return
+        }
+
         var adapters: [AnalyticsAdapter] = []
 
         // Only setup if protected data is available, the user hasn't opted out, and we aren't already registered
-        if shouldRegisterFirstPartyAdapters {
+        if !Settings.analyticsOptOut() {
             adapters = [AnalyticsLoggingAdapter(), TracksAdapter(), CrashLoggingAdapter()]
         }
 
         if FeatureFlag.podcastNewformAppsFlyer.enabled {
             adapters.append(AppsFlyerAdapter(appTrackingTransparencyProvider: AppTrackingTransparencyController.shared))
         }
+
+        if FeatureFlag.notificationsRevamp.enabled {
+            adapters.append(NotificationsCoordinator.shared)
+        }
+
         Analytics.register(adapters: adapters)
         Analytics.add(analyticsAppThemeProvider: AnalyticsAppThemeProvider())
     }