+++ podcasts/MainTabBarController.swift
@@ -104,12 +104,6 @@ class MainTabBarController: UITabBarController, NavigationProtocol {
 
         updateDatabaseIndexes()
         optimizeDatabaseIfNeeded()
-
-        if AppTrackingTransparencyController.shared.shouldShowPrompt() {
-            Task {
-                await AppTrackingTransparencyController.shared.promptConsentAlert()
-            }
-        }
     }
 
     /// Update database indexes and delete unused columns
@@ -157,7 +151,17 @@ class MainTabBarController: UITabBarController, NavigationProtocol {
             return
         }
 
-        NavigationManager.sharedManager.navigateTo(NavigationManager.onboardingFlow, data: ["flow": OnboardingFlow.Flow.initialOnboarding])
+        if AppTrackingTransparencyController.shared.shouldShowPrompt() {
+            Task {
+                await AppTrackingTransparencyController.shared.promptConsentAlert()
+            }
+        }
+
+        if FeatureFlag.encourageAccountCreation.enabled {
+            NavigationManager.sharedManager.navigateTo(NavigationManager.onboardingFlow, data: ["flow": OnboardingFlow.Flow.encourageAccountCreation])
+        } else {
+            NavigationManager.sharedManager.navigateTo(NavigationManager.onboardingFlow, data: ["flow": OnboardingFlow.Flow.initialOnboarding])
+        }
 
         // Set the flag so the user won't see the on launch flow again
         Settings.shouldShowInitialOnboardingFlow = false
@@ -332,6 +336,17 @@ class MainTabBarController: UITabBarController, NavigationProtocol {
         }
     }
 
+    func navigateToDiscover(listID: String, animated: Bool) {
+        switchToTab(.discover)
+        if let index = pcTabs.firstIndex(of: .discover),
+           let navController = viewControllers?[safe: index] as? UINavigationController {
+            navController.popToRootViewController(animated: false)
+            if let discoverDelegate = navController.topViewController as? DiscoverDelegate {
+                discoverDelegate.navigateTo(listID: listID)
+            }
+        }
+    }
+
     func navigateToUpNext(_ animated: Bool) {
         switchToTab(.upNext)
     }
@@ -340,14 +355,19 @@ class MainTabBarController: UITabBarController, NavigationProtocol {
         switchToTab(.profile)
     }
 
-    func navigateToFilter(_ filter: EpisodeFilter, animated: Bool) {
-        if !switchToTab(.filter) { return }
+    func navigateToFilter(_ filter: EpisodeFilter?, animated: Bool) {
+        guard switchToTab(.filter) else { return }
 
-        if let index = pcTabs.firstIndex(of: .filter),
-           let navController = viewControllers?[safe: index] as? UINavigationController,
-           let filtersViewController = navController.viewControllers[safe: 0] as? PlaylistsViewController {
-            filtersViewController.showFilter(filter)
+        guard let navController = selectedViewController as? UINavigationController else {
+            return
         }
+        navController.popToRootViewController(animated: false)
+
+        guard let filter,
+              let filtersViewController = navController.topViewController as? PlaylistsViewController else {
+            return
+        }
+        filtersViewController.showFilter(filter)
     }
 
     func navigateToEditFilter(_ filter: EpisodeFilter) {
@@ -449,13 +469,32 @@ class MainTabBarController: UITabBarController, NavigationProtocol {
         switchToTab(.filter)
     }
 
-    func showSettingsAppearance() {
+    func showSettings(row: SettingsViewController.TableRow?) {
+        switchToTab(.profile)
+        guard let navController = selectedViewController as? UINavigationController else { return }
+
+        navController.popViewController(animated: false)
+        let settingViewController = SettingsViewController()
+        navController.pushViewController(settingViewController, animated: row == nil)
+
+        guard let row else { return }
+
+        settingViewController.selectRow(row)
+    }
+
+    func showSettingsAppearance(showThemeSelection: Bool = false) {
         switchToTab(.profile)
         if let navController = selectedViewController as? UINavigationController {
             navController.popToRootViewController(animated: false)
 
             navController.pushViewController(SettingsViewController(), animated: false)
-            navController.pushViewController(AppearanceViewController(), animated: true)
+            let appearanceViewController = AppearanceViewController()
+            navController.pushViewController(appearanceViewController, animated: !showThemeSelection)
+            if showThemeSelection {
+                appearanceViewController.presentThemePicker(selectedTheme: Theme.preferredLightTheme()) { [weak self] theme in
+                    Theme.setPreferredLightTheme(theme, systemIsDark: self?.traitCollection.userInterfaceStyle == .dark)
+                }
+            }
         }
     }
 
@@ -511,6 +550,14 @@ class MainTabBarController: UITabBarController, NavigationProtocol {
             let generalSettingsController = GeneralSettingsViewController()
             generalSettingsController.scrollToRow = row
             navController.pushViewController(generalSettingsController, animated: true)
+        }
+    }
+
+    func showSignUp() {
+        switchToTab(.podcasts)
+        selectedViewController?.dismiss(animated: false)
+        if let controller = view.window?.rootViewController {
+            showSubscriptionRequired(controller, source: .unknown, context: nil, flow: .initialOnboarding)
         }
     }
 
@@ -643,6 +690,7 @@ class MainTabBarController: UITabBarController, NavigationProtocol {
     // MARK: - End of Year
 
     private func updateTabBarColor() {
+        self.view.backgroundColor = AppTheme.viewBackgroundColor()
         let appearance = UITabBarAppearance()
         appearance.configureWithOpaqueBackground()
         appearance.backgroundColor = AppTheme.tabBarBackgroundColor()