+++ podcasts/Onboarding/Encourage Account Creation/Banner/InformationalBannerView.swift
@@ -1,0 +1,112 @@
+import SwiftUI
+
+struct InformationalBannerView: View {
+    @EnvironmentObject var theme: Theme
+
+    @ObservedObject var viewModel: InformationalBannerViewModel
+
+    private var backgroundColor: Color {
+        if case .profile = viewModel.bannerType {
+            if case .radioactive = theme.activeTheme {
+                return theme.primaryUi06
+            }
+            return theme.primaryUi02Active
+        }
+        switch theme.activeTheme {
+        case .indigo:
+            return theme.primaryUi02Active
+        case .contrastLight:
+            return theme.secondaryUi02
+        case .contrastDark:
+            return theme.primaryUi02Active
+        default:
+            return theme.primaryUi01
+        }
+    }
+
+    private var paddingBottom: CGFloat {
+        if case .listeningHistory = viewModel.bannerType {
+            return 0
+        }
+        return 16.0
+    }
+
+    private var paddingTop: CGFloat {
+        if case .listeningHistory = viewModel.bannerType {
+            return 22.0
+        }
+        return 16.0
+    }
+
+    var body: some View {
+        HStack(alignment: .top) {
+            Image(viewModel.bannerType.iconName)
+                .accessibilityLabel(viewModel.bannerType.accessibilityLabel)
+                .foregroundColor(theme.primaryText01)
+            VStack(alignment: .leading, spacing: 6) {
+                Text(viewModel.bannerType.title)
+                    .font(.system(size: 15, weight: .medium))
+                    .foregroundColor(theme.primaryText01)
+                Text(viewModel.bannerType.description)
+                    .lineLimit(2)
+                    .fixedSize(horizontal: false, vertical: true)
+                    .multilineTextAlignment(.leading)
+                    .font(.system(size: 14))
+                    .foregroundColor(theme.primaryText02)
+                Button() {
+                    viewModel.onCreateFreeAccountTap?()
+                } label: {
+                    Text(L10n.eacInformationalBannerCreateAccount)
+                        .font(.system(size: 14, weight: .medium))
+                        .foregroundColor(theme.primaryText02Selected)
+                }
+                .padding(.top, 6.0)
+            }
+            Spacer()
+        }
+        .padding(.top, 20.0)
+        .padding(.leading, 16.0)
+        .padding(.trailing, 56.0)
+        .padding(.bottom, 20.0)
+        .background(backgroundColor)
+        .cornerRadius(8)
+        .overlay(alignment: .topTrailing) {
+            Button() {
+                viewModel.onCloseBannerTap?()
+            } label: {
+                Image("close")
+                    .renderingMode(.template)
+                    .foregroundColor(theme.primaryText01)
+            }
+            .padding(.top, 16.0)
+            .padding(.trailing, 16.0)
+        }
+        .padding(.bottom, paddingBottom)
+        .padding(.top, paddingTop)
+        .padding(.horizontal, 16.0)
+    }
+}
+
+struct InformationalBannerViewFilters_Preview: PreviewProvider {
+    static var previews: some View {
+        VStack {
+            Group {
+                InformationalBannerView(viewModel: InformationalBannerViewModel(bannerType: .filters))
+                    .environmentObject(Theme(previewTheme: .light))
+                    .previewLayout(.sizeThatFits)
+                    .frame(width: 386, height: 170)
+
+                InformationalBannerView(viewModel: InformationalBannerViewModel(bannerType: .listeningHistory))
+                    .environmentObject(Theme(previewTheme: .light))
+                    .previewLayout(.sizeThatFits)
+                    .frame(width: 386, height: 170)
+            }
+            .background(.gray.opacity(0.5))
+
+            InformationalBannerView(viewModel: InformationalBannerViewModel(bannerType: .profile))
+                .environmentObject(Theme(previewTheme: .light))
+                .previewLayout(.sizeThatFits)
+                .frame(width: 386, height: 170)
+        }
+    }
+}