+++ podcasts/WatchManager.swift
@@ -12,6 +12,10 @@ class WatchManager: NSObject, WCSessionDelegate {
     // The last retrieved log is cached here for the duration of this session
     var cachedLog: String? = nil
 
+    var isWatchAppInstalled: Bool {
+        return WCSession.isSupported() && WCSession.default.isWatchAppInstalled
+    }
+
     func setup() {
         if !WCSession.isSupported() { return }
 
@@ -34,6 +38,8 @@ class WatchManager: NSObject, WCSessionDelegate {
         NotificationCenter.default.addObserver(self, selector: #selector(autoDownloadChanged), name: Constants.Notifications.watchAutoDownloadSettingsChanged, object: nil)
         NotificationCenter.default.addObserver(self, selector: #selector(playbackStateChanged), name: Constants.Notifications.podcastChapterChanged, object: nil)
         NotificationCenter.default.addObserver(self, selector: #selector(playbackStateChanged), name: Constants.Notifications.podcastChaptersDidUpdate, object: nil)
+
+        cachedLog = WatchManager.shared.readLogFile()
     }
 
     deinit {
@@ -394,6 +400,13 @@ class WatchManager: NSObject, WCSessionDelegate {
         }
         DispatchQueue.global(qos: .background).async { [weak self] in
             self?.sendStateToWatch()
+            if FeatureFlag.refreshAndSaveWatchLogsOnSend.enabled {
+                FileLog.shared.addMessage("WatchManager: Start log request on send")
+                WatchManager.shared.requestLogFile(completion: { _ in
+                    // We don't do anything with the log, requesting it will cache and save the contents
+                    FileLog.shared.addMessage("WatchManager: Completed log request on send")
+                })
+            }
         }
     }