+++ CHANGELOG.md
@@ -1,12 +1,18 @@
-7.87
+7.88
 -----
 
+
+7.87
+-----
+- Add Encourage Account Creation feature [#2905](https://github.com/Automattic/pocket-casts-ios/issues/2905)
 
 7.86
 -----
 - New podcast page design [#2825](https://github.com/Automattic/pocket-casts-ios/pull/2825)
 - Add Creator Funding Support [#2913](https://github.com/Automattic/pocket-casts-ios/pull/2913)
 - Fix Watch app should pause play during announcements from other apps [#2245](https://github.com/Automattic/pocket-casts-ios/issues/2245)
+- Fix download indicator color in the up next list [#2941](https://github.com/Automattic/pocket-casts-ios/pull/2941)
+- Fix flashing when switching tabs [#2964](https://github.com/Automattic/pocket-casts-ios/pull/2964)
 
 7.85
 -----